requirements.txt
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0
requests>=2.32,<3.0
pytest>=8.0,<9.0
